requirements.txt
numpy
cupy-cuda11x
pillow
open3d
tqdm
scipy
scikit-learn